The Injector Part
The Injector : - receive the orders form the servers, match this orders
with the portfolio of the client, send the orders to the journal table
for each client thanks to cron-populate.php. Becarefull to add security
settings from IP in server production if you use it LIVE ! ..the
security are not present inside this scripts which are given without the
security layer.